Changed: unified config.js file for tests, utils.js file for token tests, refactored package.json, small changes after review:
