+++ contracts/token/ERC20ForSpl/README.md
@@ -120,7 +120,7 @@ by the `balanceOf` function and this ATA token balance will be spendable via the
 ## Running tests
 * **ERC20ForSpl** - this action is done in 2 steps. First we have to deploy a SPLToken on Solana and then to pass it into the ERC20ForSpl's constructor.
     * run ```node test/helpers/deploySPLToken.js``` to deploy new SPLToken.
-    * paste the new SPLToken mint address to `config.DATA.ADDRESSES.ERC20ForSplTokenMint` inside `test/config.js`.
+    * paste the new SPLToken mint address to `utils.DATA.ADDRESSES.ERC20ForSplTokenMint` inside `test/utils.js`.
     * run `npx hardhat test test/ERC20ForSpl.js --network neondevnet`
 * **ERC20ForSplMintable** - run `npx hardhat test test/ERC20ForSplMintable.js --network neondevnet`
