+++ package.json
@@ -1,14 +1,21 @@
 {
-  "name": "neon-tutorials-hardhat",
+  "name": "neon-contracts",
   "version": "1.0.0",
   "description": "",
   "main": "index.js",
   "scripts": {
-    "test-composability-curvestand": "npx hardhat test ./tests/composability/system.test.js --network curvestand && npx hardhat test ./tests/composability/spl-token.test.js --network curvestand"
+    "test-composability-curvestand": "npx hardhat test ./test/composability/system.test.js --network curvestand && npx hardhat test ./test/composability/spl-token.test.js --network curvestand",
+    "test-composability-neondevnet": "npx hardhat test ./test/composability/system.test.js --network neondevnet && npx hardhat test ./test/composability/spl-token.test.js --network neondevnet",
+    "test-raydium-composability-neondevnet": "npx hardhat test ./test/composability/raydium.test.js --network neondevnet && npx hardhat test ./test/composability/raydium-create-pool-and-lock-LP.test.js --network neondevnet",
+    "test-token-curvestand": "npx hardhat test ./test/token/ERC20ForSplFactory.test.js --network curvestand && npx hardhat test ./test/token/ERC20ForSPL.test.js --network curvestand && npx hardhat test ./test/token/ERC20ForSPLMintable.test.js --network curvestand",
+    "test-token-neondevnet": "npx hardhat test ./test/token/ERC20ForSplFactory.test.js --network neondevnet && npx hardhat test ./test/token/ERC20ForSPL.test.js --network neondevnet && npx hardhat test ./test/token/ERC20ForSPLMintable.test.js --network neondevnet"
   },
   "author": "",
   "license": "ISC",
-  "devDependencies": {
+  "dependencies": {
+    "@metaplex-foundation/js": "^0.20.1",
+    "@neonevm/token-transfer-core": "^4.4.2",
+    "@neonevm/token-transfer-ethers": "^4.4.2",
     "@nomicfoundation/hardhat-chai-matchers": "^2.0.6",
     "@nomicfoundation/hardhat-ethers": "^3.0.6",
     "@nomicfoundation/hardhat-network-helpers": "^1.0.10",
@@ -17,6 +24,8 @@
     "@openzeppelin/contracts": "5.2.0",
     "@openzeppelin/contracts-upgradeable": "5.2.0",
     "@openzeppelin/hardhat-upgrades": "^3.5.0",
+    "@solana/spl-token": "^0.4.12",
+    "@solana/web3.js": "^1.98.0",
     "@typechain/ethers-v6": "^0.5.1",
     "@typechain/hardhat": "^9.1.0",
     "@types/chai": "^4.3.16",
@@ -27,16 +36,9 @@
     "hardhat": "^2.22.6",
     "hardhat-gas-reporter": "^1.0.10",
     "solidity-coverage": "^0.8.13",
+    "solidity-docgen": "^0.6.0-beta.36",
     "ts-node": "^10.9.2",
     "typechain": "^8.3.2",
     "typescript": "^5.4.5"
-  },
-  "dependencies": {
-    "@metaplex-foundation/js": "^0.20.1",
-    "@neonevm/token-transfer-core": "^4.4.2",
-    "@neonevm/token-transfer-ethers": "^4.4.2",
-    "@solana/spl-token": "^0.4.12",
-    "@solana/web3.js": "^1.98.0",
-    "solidity-docgen": "^0.6.0-beta.36"
   }
 }
