+++ test/token/helpers/deploySPLToken.js
@@ -12,7 +12,8 @@ const {
 const { Metaplex } = require("@metaplex-foundation/js");
 const bs58 = require("bs58");
 const { createCreateMetadataAccountV3Instruction } = require("@metaplex-foundation/mpl-token-metadata");
-const { config } = require('../config');
+const utils = require('../utils');
+const config = require('../../config.js');
 require("dotenv").config({path: __dirname + '/../../.env'});
 
 const connection = new web3.Connection(config.svm_node[network.name], "processed");
@@ -28,7 +29,7 @@ const solanaUser4 = web3.Keypair.fromSecretKey( // Solana user with tokens balan
 
 async function init() {
     if (await connection.getBalance(keypair.publicKey) == 0) {
-        await config.utils.airdropSOL(keypair);
+        await utils.airdropSOL(keypair);
     }
 
     const seed = 'seed' + Date.now().toString(); // random seed on each script call
